# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: TAIFA Platform API
  version: 1.0.0
  description: Authoritative TAIFA Platform APIs for identity-bound access, payments
    over the double-entry ledger, enterprise financial operations, commerce and national
    mobility.
paths:
...
  /api/v1/ecosystem/domains:
    get:
      operationId: api_v1_ecosystem_domains_retrieve
      tags:
      - platform
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
...
  /api/v1/ecosystem/webhooks:
    get:
      operationId: api_v1_ecosystem_webhooks_retrieve
      tags:
      - platform-open
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
    post:
      operationId: api_v1_ecosystem_webhooks_create
      tags:
      - platform-open
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    DeviceToken:
      type: http
      scheme: bearer
      description: 'Device-bound token from `POST /auth/device/register`. Send it
        as `Authorization: Bearer <token>` together with the `X-Device-Id` header.'
servers:
- url: /
  description: current host
...